backend/requirements.txt
fastapi
mysqlclient
uvicorn
sqlalchemy
pymysql
python-multipart
PyJWT
cryptography 